RFC: Simpler array hashing
This is a straw-man implementation of a simpler array hashing scheme.  It's very basic and has room for further optimizations, but it's intention is to provide a starting point as an alternative to #25822.  In short: This hashes the size of the array and then the first three distinct elements and their linear indices and then the last three distinct elements and their linear distance from the end of the array.

The exact scheme here is open to bikeshedding -- we could use more or less distinct elements.  I use "distinct elements" as a way of ensuring that all relatively empty sparse arrays don't hash similarly, and I hash elements at both the beginning and end of the array because they're the simplest to get to and final elements will be the "most expensive" to discover that they differ if we have to fall back to an equality check.