Fix handling of missing values in median and quantile and improve docs
Ensure we always return missing or throw an error in the presence of missing values.
Support skipmissing(v) by supporting any iterator and calling collect on it, which
makes median and quantile consistent with mean and sum.
Improve docstrings for stats functions by adding examples and making them more consistent.